requirements.txt
# Core dependencies
gradio>=4.0.0
openai>=1.0.0
tiktoken>=0.5.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
simplemma>=0.9.1

# Document processing
pymupdf>=1.22.0  # PDF processing
python-docx>=0.8.11  # DOCX processing

# Excel processing (optional, needed only for Excel translation)
xlwings>=0.30.0  # Excel processing (requires Microsoft Excel installation)

# Suggested packages for improved functionality
icecream>=2.1.3  # Better debugging 
reportlab